Implement Redis lock key generation for service transitions
